# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateTableRequest:
    properties:
      fields:
        items:
          $ref: '#/definitions/models.FieldRequest'
        minItems: 1
        type: array
      id:
        example: users_table
        type: string
    required:
    - fields
    - id
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  models.FieldChangeResponse:
    properties:
      entity_id:
        example: entity123
        type: string
      new_value:
        example: new_name
        type: string
      old_value:
        example: old_name
        type: string
      timestamp:
        example: "2023-12-31T23:59:59Z"
        type: string
    type: object
  models.FieldHistoryResponse:
    properties:
      changes:
        items:
          $ref: '#/definitions/models.FieldChangeResponse'
        type: array
      field_name:
        example: user_name
        type: string
      table_id:
        example: users_table
        type: string
    type: object
  models.FieldRequest:
    properties:
      data_type:
        enum:
        - string
        - int
        - float
        - bool
        - date
        - json
        - reference
        example: string
        type: string
      name:
        example: user_name
        type: string
    required:
    - data_type
    - name
    type: object
  models.FieldResponse:
    properties:
      data_type:
        example: string
        type: string
      name:
        example: user_name
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
  models.TableHistoryResponse:
    properties:
      changes:
        items:
          $ref: '#/definitions/models.Tuple'
        type: array
      table_id:
        example: users_table
        type: string
    type: object
  models.TableListResponse:
    properties:
      tables:
        items:
          $ref: '#/definitions/models.TableResponse'
        type: array
    type: object
  models.TableResponse:
    properties:
      created_at:
        example: "2023-12-31T23:59:59Z"
        type: string
      fields:
        items:
          $ref: '#/definitions/models.FieldResponse'
        type: array
      id:
        example: users_table
        type: string
      updated_at:
        example: "2023-12-31T23:59:59Z"
        type: string
      user_id:
        example: user123
        type: string
    type: object
  models.Tuple:
    properties:
      entity_id:
        type: string
      field_name:
        type: string
      gsi1_pk:
        description: GSI keys for different query patterns
        type: string
      gsi1_sk:
        description: '{timestamp}#{entityID}'
        type: string
      gsi2_pk:
        description: USER#{userID}#{tableID}#{entityID}
        type: string
      gsi2_sk:
        description: '{timestamp}#{fieldName}'
        type: string
      gsi3_pk:
        description: USER#{userID}#{tableID}
        type: string
      gsi3_sk:
        description: '{entityID}#{timestamp}#{fieldName}'
        type: string
      pk:
        description: DynamoDB keys
        type: string
      sk:
        description: 'Sort key: TUPLE#{tableID}#{entityID}#{timestamp}#{fieldName}'
        type: string
      table_id:
        type: string
      timestamp:
        type: string
      user_id:
        type: string
      value:
        type: string
    type: object
...
  models.UpdateTableRequest:
    properties:
      fields:
        items:
          $ref: '#/definitions/models.FieldRequest'
        minItems: 1
        type: array
    required:
    - fields
    type: object
  models.UserInfoResponse:
    properties:
      created_at:
        example: "2023-12-31T23:59:59Z"
        type: string
      updated_at:
        example: "2023-12-31T23:59:59Z"
        type: string
      user_id:
        example: user123
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Notably API server for managing time-series data and entities.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Notably API
  version: "1.0"
paths:
...
  /tables:
    get:
      consumes:
      - application/json
      description: Get a list of all tables for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of tables
          schema:
            $ref: '#/definitions/models.TableListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: List all tables
      tags:
      - tables
    post:
      consumes:
      - application/json
      description: Create a new table with field definitions
      parameters:
      - description: Table creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateTableRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Table created successfully
          schema:
            $ref: '#/definitions/models.TableResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new table
      tags:
      - tables
  /tables/{tableId}:
    delete:
      consumes:
      - application/json
      description: Delete a table by its ID
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Table deleted successfully
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete table
      tags:
      - tables
    get:
      consumes:
      - application/json
      description: Retrieve a specific table by its ID
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Table information
          schema:
            $ref: '#/definitions/models.TableResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Table not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get table by ID
      tags:
      - tables
    put:
      consumes:
      - application/json
      description: Update table field definitions
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Table update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Table updated successfully
          schema:
            $ref: '#/definitions/models.TableResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Table not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update table
      tags:
      - tables
...
  /tables/{tableId}/history:
    get:
      consumes:
      - application/json
      description: Retrieve the history of changes for a specific table
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Start date for history (RFC3339 format)
        in: query
        name: since
        type: string
      - description: Maximum number of history entries to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Table history
          schema:
            $ref: '#/definitions/models.TableHistoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get table history
      tags:
      - tables
  /tables/{tableId}/history/fields/{fieldName}:
    get:
      consumes:
      - application/json
      description: Retrieve the history of changes for a specific field in a table
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Field name
        in: path
        name: fieldName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Field history
          schema:
            $ref: '#/definitions/models.FieldHistoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get field history
      tags:
      - tables
  /users/me:
    get:
      consumes:
      - application/json
      description: Retrieve the current authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User information
          schema:
            $ref: '#/definitions/models.UserInfoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get current user information
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the current authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "501":
          description: Not implemented
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update current user profile
      tags:
      - users
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"